fix: moving models
