+++ package/cache/main.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"itemmeli/models"
 	"itemmeli/package/config"
-	"itemmeli/package/models"
 	"time"
 
 	"github.com/redis/go-redis/v9"
