+++ package/database/main.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"itemmeli/models"
 	"itemmeli/package/config"
-	"itemmeli/package/models"
 	"os"
 )
